+++ CHANGES.rst
@@ -5,6 +5,8 @@ Changelog
 ------------------
 
 - switched to `kasperl` library for base API and generic pipeline plugins
+- morphological filters now have an option to specify the output format
+  (as-is/binary/grayscale/rgb)
 
 
 0.0.1 (2025-07-11)